add permissions authentication on board

+++ backend/controllers/board.js
@@ -58,6 +58,9 @@ const addMember = async (req, res) => {
   let member = await Board.findById(req.body.boardId);
   if (!member) return res.status(400).send("Board doesn't exist");
 
+  if (member.userId.toString() !== req.user._id.toString())
+    return res.status(400).send("You have no permisiion");
+
   let newMember = member.members;
   let data = {
     id: user._id,
@@ -93,6 +96,9 @@ const deleteMember = async (req, res) => {
   let member = await Board.findById(req.body.boardId);
   if (!member) return res.status(400).send("Board doesn't exist");
 
+  if (member.userId.toString() !== req.user._id.toString())
+    return res.status(400).send("You have no permisiion");
+
   let delMember = member.members;
 
   for (var i = 0; i < delMember.length; i++) {
@@ -112,17 +118,17 @@ const deleteMember = async (req, res) => {
 };
 
 const listBoard = async (req, res) => {
-  let board = await Board.find({userId: req.user._id});
+  let board = await Board.find({ userId: req.user._id });
   if (!board || board.length === 0)
     return res.status(400).send("You have no assigned tasks");
   return res.status(200).send({ board });
 };
 
 const listBoardMember = async (req, res) => {
-  let user  = await User.findById(req.user._id);
-  if(!user) return res.status(400).send("User not found");
+  let user = await User.findById(req.user._id);
+  if (!user) return res.status(400).send("User not found");
 
-  let board = await Board.find({'members.id' : user._id });
+  let board = await Board.find({ "members.id": user._id });
   if (!board || board.length === 0)
     return res.status(400).send("You have no assigned tasks");
   return res.status(200).send({ board });
@@ -132,19 +138,46 @@ const deleteBoard = async (req, res) => {
   let validId = mongoose.Types.ObjectId.isValid(req.params._id);
   if (!validId) return res.status(400).send("Invalid id");
 
-  let taskImg = await Board.findById(req.params._id);
-  taskImg = taskImg.imageUrl;
-  taskImg = taskImg.split("/")[4];
-  let serverImg = "./uploads/" + taskImg;
+  let boardImg = await Board.findById(req.params._id);
+  if (!boardImg) return res.status(400).send("Board not found");
+
+  if (boardImg.userId.toString() !== req.user._id.toString())
+    return res.status(400).send("You have no permisiion");
+
+  boardImg = boardImg.imageUrl;
+  boardImg = boardImg.split("/")[4];
+  let serverImg = "./uploads/" + boardImg;
 
   let board = await Board.findByIdAndDelete(req.params._id);
   if (!board) return res.status(400).send("Board not found");
+
   try {
     fs.unlinkSync(serverImg);
   } catch (error) {
     console.log("Image no found in server");
   }
   return res.status(200).send({ message: "deleted board" });
+};
+
+const updateBoard = async (req, res) => {
+  let validId = mongoose.Types.ObjectId.isValid(req.body._id);
+  if (!validId) return res.status(400).send("Invalid id");
+
+  if (!req.body.name || !req.body.description)
+    return res.status(400).send("Incomplete data");
+
+  const memberInfo = await Board.findById(req.body._id);
+  if (!memberInfo) return res.status(400).send("Board not found");
+
+  if (memberInfo.userId.toString() !== req.user._id.toString())
+    return res.status(400).send("You have no permisiion");
+
+  const board = await Board.findByIdAndUpdate(req.body._id, {
+    name: req.body.name,
+    description: req.body.description,
+  });
+  if (!board) return res.status(400).send("Board not found");
+  return res.status(200).send({ board });
 };
 
 module.exports = {
@@ -154,4 +187,5 @@ module.exports = {
   addMember,
   deleteMember,
   deleteBoard,
+  updateBoard,
 };
